interface/requirements.txt
allure-pytest
appium-python-client
pytest
pytest-testconfig
requests
slenium
urllib3
